requirements.txt
fastmcp>=2.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
cachetools>=5.3.0